FEATURE: Pass the task entity to the task implementation
This change passes the task entity as a constructor argument to
the task implementation. When the task implementation does not
have a constructor this is completely ignored. If it has one it allows
for changing the status of the task.

This can be used to disable a job automatically after execution allowing
for manual execution only.